reposition dashboard
